readability renaming
gave some files more meaningful names
