+++ templates/index.html
@@ -2,7 +2,7 @@
 <html>
 <head>
     <title>NAS UI</title>
-    <script src="/static/script.js" defer></script>
+    <script src="/static/directory.js" defer></script>
     <link rel="stylesheet" href="/static/styles.css">
 </head>
 <body>
